+++ packages/sources/bank-frick/src/endpoint/index.ts
@@ -1,5 +1,1 @@
-import type { TInputParameters as AccountInputParameters } from './accounts'
-
-export type TInputParameters = AccountInputParameters
-
-export * as accounts from './accounts'
+export { accountsRestEndpoint as accounts } from './accounts'