+++ packages/sources/coingecko-test/src/endpoint/dominance.ts
@@ -1,5 +1,4 @@
 import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
-import { HttpRequestConfig } from '@chainlink/external-adapter-framework/transports'
 import { BatchWarmingTransport } from '@chainlink/external-adapter-framework/transports/batch-warming'
 import {
   buildGlobalRequestBody,
@@ -9,9 +8,7 @@ import {
 } from '../global-utils'
 
 const batchEndpointTransport = new BatchWarmingTransport<GlobalEndpointTypes>({
-  prepareRequest: (_, config): HttpRequestConfig<never> => {
-    return buildGlobalRequestBody(config.API_KEY)
-  },
+  prepareRequest: (_, config) => buildGlobalRequestBody(config),
   parseResponse: (params, res) =>
     params.map((requestPayload) => constructEntry(res, requestPayload, 'market_cap_percentage')),
 })