+++ packages/sources/dxfeed-secondary/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/dxfeed-secondary-adapter
 
+## 1.2.26
+
+### Patch Changes
+
+- Updated dependencies [0719f739b]
+  - @chainlink/ea-bootstrap@2.25.1
+  - @chainlink/ea-test-helpers@1.4.1
+  - @chainlink/dxfeed-adapter@1.3.26
+
 ## 1.2.25
 
 ### Patch Changes